Refactor complete and bugs fixed
