+++ client/src/api/agentsAPI.ts
@@ -1,4 +1,5 @@
-const BASE_URL = `${import.meta.env.VITE_API_BASE_URL}/cookie-fun`;
+const BASE_URL = "http://localhost:8080/api"; // Your Flask API base URL
+const BASE_URL_COOKIE_FUN = `${import.meta.env.VITE_API_BASE_URL}/cookie-fun`;
 
 // Function to get all agents
 export async function getAgents() {
@@ -190,50 +191,50 @@ export async function updateSeasons(agentName: string, seasons: any[]) {
   return await response.json();
 }
 
-// Add these new interfaces
-interface CookieFunResponse {
-  [key: string]: any;
-}
+// // Add these new interfaces
+// interface CookieFunResponse {
+//   [key: string]: any;
+// }
 
-interface PaginationParams {
-  page?: number;
-  pageSize?: number;
-  interval?: string;
-}
+// interface PaginationParams {
+//   page?: number;
+//   pageSize?: number;
+//   interval?: string;
+// }
 
-// Add these new functions
-export async function getCookieFunAgentsPaged(params: PaginationParams = {}): Promise<CookieFunResponse> {
-  const queryParams = new URLSearchParams({
-    page: (params.page || 1).toString(),
-    pageSize: (params.pageSize || 25).toString(),
-    interval: params.interval || '_7Days'
-  });
+// // Add these new functions
+// export async function getCookieFunAgentsPaged(params: PaginationParams = {}): Promise<CookieFunResponse> {
+//   const queryParams = new URLSearchParams({
+//     page: (params.page || 1).toString(),
+//     pageSize: (params.pageSize || 25).toString(),
+//     interval: params.interval || '_7Days'
+//   });
 
-  const response = await fetch(`${BASE_URL}/agents/paged?${queryParams}`);
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-  return await response.json();
-}
+//   const response = await fetch(`${BASE_URL}/agents/paged?${queryParams}`);
+//   if (!response.ok) {
+//     throw new Error(`HTTP error! status: ${response.status}`);
+//   }
+//   return await response.json();
+// }
 
-export async function getCookieFunAgentByTwitter(
-  username: string, 
-  interval: string = '_7Days'
-): Promise<CookieFunResponse> {
-  const response = await fetch(`${BASE_URL}/twitter/${username}?interval=${interval}`);
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-  return await response.json();
-}
+// export async function getCookieFunAgentByTwitter(
+//   username: string, 
+//   interval: string = '_7Days'
+// ): Promise<CookieFunResponse> {
+//   const response = await fetch(`${BASE_URL}/twitter/${username}?interval=${interval}`);
+//   if (!response.ok) {
+//     throw new Error(`HTTP error! status: ${response.status}`);
+//   }
+//   return await response.json();
+// }
 
-export async function getCookieFunAgentByContract(
-  address: string,
-  interval: string = '_7Days'
-): Promise<CookieFunResponse> {
-  const response = await fetch(`${BASE_URL}/contract/${address}?interval=${interval}`);
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-  return await response.json();
-}
+// export async function getCookieFunAgentByContract(
+//   address: string,
+//   interval: string = '_7Days'
+// ): Promise<CookieFunResponse> {
+//   const response = await fetch(`${BASE_URL_COOKIE_FUN}/contract/${address}?interval=${interval}`);
+//   if (!response.ok) {
+//     throw new Error(`HTTP error! status: ${response.status}`);
+//   }
+//   return await response.json();
+// }
