+++ client/src/interfaces/cookie.FunInterfaces.ts
@@ -1,0 +1,10 @@
+// Add these new interfaces
+export interface CookieFunResponse {
+    [key: string]: any;
+}
+  
+export interface PaginationParams {
+    page?: number;
+    pageSize?: number;
+    interval?: string;
+}
