+++ server_node/src/server.ts
@@ -3,7 +3,7 @@ import path from 'path';
 import cors from 'cors';
 
 // Load env variables before other imports, pointing to the root .env file
-dotenv.config({ path: path.resolve(__dirname, '../.env') });
+dotenv.config({ path: path.resolve(__dirname, '../../.env') });
 
 import express from 'express';
 import { router as cookieFunRoutes } from './routes/cookieDotFunRoutes';
@@ -23,8 +23,8 @@ app.use('/api/cookie-fun', cookieFunRoutes);
 // Add future API routes here
 // app.use('/api/other-service', otherApiRoutes);
 
-// Error handling middleware
-app.use((err: any, req: express.Request, res: express.Response, next: express.NextFunction) => {
+// Error handling middleware - Fixed unused parameters warning
+app.use((err: any, _req: express.Request, res: express.Response, _next: express.NextFunction) => {
     console.error(err.stack);
     res.status(500).json({ error: 'Something went wrong!' });
 });
